Update template-sync.yml (#5)
## 1) Description
all rights.
Removed workflows from the updated files. 

## 2) Technical choice

One day there will be documentation for this. Good documentation. 
Updating the workflows requires specific rights that can only be given
through a personal access token. Configuring this is one more step and
will thus be made opt-in and described in the documentation.