Refactor conditional rendering on preview buttons
Conditional rendering on preview buttons was changed, because due to high number of options in mode editor and preview, this involve an out of screen of options. To avoid this issue I regrouped lonely options in a single button which open a menu. Moreover, this commit hold a new function to detect the position of cursor in the textarea (unused for the moment).